Moved IBotControllerBehavior to TroopSpawner extension.
