+++ Alliance.Client/Extensions/TroopSpawner/Handlers/SpawnTroopHandler.cs
@@ -1,7 +1,7 @@
 ﻿using Alliance.Client.Extensions.TroopSpawner.Models;
-using Alliance.Common.Core;
 using Alliance.Common.Core.ExtendedCharacter.Extension;
 using Alliance.Common.Extensions;
+using Alliance.Common.Extensions.TroopSpawner.Interfaces;
 using Alliance.Common.Extensions.TroopSpawner.Models;
 using Alliance.Common.Extensions.TroopSpawner.NetworkMessages.FromServer;
 using NetworkMessages.FromServer;
