+++ Alliance.Common/GameModes/PvC/Behaviors/PvCGameModeClientBehavior.cs
@@ -1,6 +1,6 @@
-﻿using Alliance.Common.Core;
-using Alliance.Common.Core.Configuration.Models;
+﻿using Alliance.Common.Core.Configuration.Models;
 using Alliance.Common.Core.ExtendedCharacter.Models;
+using Alliance.Common.Extensions.TroopSpawner.Interfaces;
 using Alliance.Common.Extensions.TroopSpawner.Models;
 using System.Linq;
 using System.Reflection;
@@ -20,7 +20,7 @@ namespace Alliance.Common.GameModes.PvC.Behaviors
         {
         }
 
-        public override bool IsGameModeUsingGold => true;
+        public override bool IsGameModeUsingGold => false;
 
         public override void OnBehaviorInitialize()
         {
