+++ Alliance.Common/GameModes/Siege/Behaviors/PvCMissionMultiplayerSiegeClient.cs
@@ -1,4 +1,4 @@
-﻿using Alliance.Common.Core;
+﻿using Alliance.Common.Extensions.TroopSpawner.Interfaces;
 using TaleWorlds.MountAndBlade;
 
 namespace Alliance.Common.GameModes.Siege.Behaviors
